fixes
git-svn-id: https://svn.java.net/svn/electric~svn/trunk@650 b64ed6c1-75a8-e629-b1da-d0dae74a743a

+++ srcj/com/sun/electric/database/variable/TextDescriptor.java
@@ -1070,7 +1070,7 @@ public class TextDescriptor
 		int offset = (descriptor0 & VTXOFF) >> VTXOFFSH;
 		if ((descriptor0&VTXOFFNEG) != 0) offset = -offset;
 		int scale = getOffScale() + 1;
-		return(offset * scale / 4);
+		return((double)offset * scale / 4);
 	}
 
 	/**
@@ -1082,7 +1082,7 @@ public class TextDescriptor
 		int offset = (descriptor0 & VTYOFF) >> VTYOFFSH;
 		if ((descriptor0&VTYOFFNEG) != 0) offset = -offset;
 		int scale = getOffScale() + 1;
-		return(offset * scale / 4);
+		return((double)offset * scale / 4);
 	}
 
 	/**
